packets.c: add packet type to dump a section of memory
The server can now request a section of memory to be dumped from the client.
The client will then send the requested memory section to the server, with a size
of 0xEC bytes. The default size of a Pokémon, as the upload is performed through a
currently unmodified Pokémon upload function.

The size will be be made configurable in a future commit.

Signed-off-by: Jorik Devreese <jorikdevreese@hotmail.be>